chore(sims): swap 96.sim.ts and 97.sim.ts to match on-chain proposal IDs
The proposals were posted in the opposite order from how the sim files
were numbered. Proposal 96 is the BNB/Polygon/Celo fee expansion, and
proposal 97 is the Franchiser undelegation that returns ~12.5M UNI to
the Timelock. This swaps the file contents (and docblock headers) so
SIM_NAME=96 / SIM_NAME=97 match the actual on-chain proposal numbers.
